Drop unnecessary &mut refs.
main() creates the left/right vars and takes a mutable reference
to them, for passing to join(). The &mut here is required to match
the type signature of join().

Everything from join() onwards does not need to modify the left/
right vars itself, only their contents, so there is no need to pass
a mutable reference. Or possibly it's already a mutable reference so
there is no need to 're-mut' it. Not sure.